This is an updated version of all files. Deletion of one.
I added a lot of new features to the game. It now has a game over screen
and a winning screen. At this very moment the screen size is too small
for some of the buttons. This can be changed later as to where they are
even placed. This can get more generic as all the screens with buttons
only have 3. A more OOP approach needs to be taken to make this code
much more readable and generic.