Added images to replace the background and the bricks.
This change gives the game a little more life. I originally attempted to
put a gradient in manually. I took this code from a processing forum. It
put too much strain on the game and slowed it considerably. I decided
images would be the best option to counter this issue. `